fix: prevent exception on missing output dir
Checks if the output directory exists before attempting to remove it to avoid throwing an exception.

+++ src/main/java/Main.java
@@ -61,6 +61,11 @@ public class Main
     private static void confirmAndClearDir(Path dir)
             throws IOException
     {
+        if (!Files.exists(dir))
+        {
+            return;
+        }
+
         deleteDirectoryRecursively(dir);
         System.out.println("Directory deleted.");
     }
